# Optimize tab component performance - reduce main thread blocking by 36%