--- a/DEPLOYMENT_VERIFICATION.md
+++ b/DEPLOYMENT_VERIFICATION.md
@@ -1,0 +1,1 @@
+Deployment verified successfully at: https://snap-clone-dev-754683067800.us-central1.run.app