--- a/client/src/App.css
+++ b/client/src/App.css
@@ -405,8 +405,8 @@ body {
 
 .content-grid {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
-  gap: 1.5rem;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 0.75rem;
   opacity: 1;
   transition: opacity 0.2s ease-in-out;
   contain: layout style;
@@ -418,31 +418,49 @@ body {
 }
 
 .content-tile {
-  background-color: #ffffff;
-  border-radius: 12px;
+  background-color: transparent;
+  border-radius: 8px;
   overflow: hidden;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  transition: transform 0.2s, box-shadow 0.2s;
+  position: relative;
+  transition: transform 0.2s;
   contain: layout style paint;
   content-visibility: auto;
 }
 
 .content-tile:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  transform: scale(1.02);
 }
 
 .tile-image {
   width: 100%;
-  height: 200px;
+  height: 280px;
   object-fit: cover;
+  border-radius: 8px;
   /* Performance optimizations */
   contain: layout style paint;
-  aspect-ratio: 16/9;
+  aspect-ratio: 9/16;
 }
 
 .tile-content {
   padding: 1rem;
+}
+
+/* Snapchat-style overlay for video tiles */
+.tile-overlay {
+  position: absolute;
+  bottom: 8px;
+  right: 8px;
+  z-index: 2;
+}
+
+.view-count {
+  background: rgba(0, 0, 0, 0.6);
+  color: white;
+  padding: 4px 8px;
+  border-radius: 12px;
+  font-size: 0.75rem;
+  font-weight: 500;
+  backdrop-filter: blur(4px);
 }
 
 .tile-user {