… client/lighthouse-final-report.json
… client/lighthouse-optimized-report.json
… client/lighthouse-report.json
… client/lighthouse-seo-verification.json